chore(samples): simplify and update hello world v2 sample
