+++ samples/v2/.gitignore
@@ -1,0 +1,1 @@
+*.json
