--- samples/v2/hello_world/hw.yaml
@@ -1,38 +1,0 @@
-# Copyright 2021 The Kubeflow Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-name: Hello world
-inputs:
-- {name: text, type: String}
-implementation:
-  container:
-    image: python:3.7
-    command:
-    - python3
-    - -u
-    - -c
-    - |
-      def hello_world(text):
-          print(text)
-          return text
-
-      import argparse
-      _parser = argparse.ArgumentParser(prog='Hello world', description='')
-      _parser.add_argument("--text", dest="text", type=str, required=True, default=argparse.SUPPRESS)
-      _parsed_args = vars(_parser.parse_args())
-
-      _outputs = hello_world(**_parsed_args)
-    args:
-    - --text
-    - {inputValue: text}
